Doing crossvalidation serially in a single process
The overhead of serialising data to worker processes is too great for
it to be worth it. It might be possible to reduce that overhead, but it
is easier to run in a single process. Experiments are already quite
fast, and being single core I can run more of them at a time.

See https://www.pivotaltracker.com/story/show/76271838