keeping track of repeated items to avoid writing them out
This only has a significant effect for the term counts stats file, which
contains a list of all features, incl unigrams (which tend to repeat).
The list of replacements hardly has any replacements, so this slightly
increases the size of the file. To keep the code simple I write both
files with the same function