Changed semantics of `rep` param of word2vec builder script.
Each repetition takes a random sample of N percent of the corpus and
trains the model on that. The first random sample is not actually
random, but is just consecutive files. This is so that two samples, e.g.
50% and 60% differ just by 10% of data, so we can measure how much of a
difference adding these extra 10% makes. Comparing the results of
multiple repetitions over a random sample of the data shows how much the
particular sample matters.