added ability to "equalise" coverage of thesauri
Can now only keep those vectors from thes A which are also contained in
thes B. This way we can pull apart the contribution of vector quality vs
vector quantity. There is currently some evidence suggesting having a
poor vector for a NP is better than not having a vector at all. This is
not what we want the framework to measure