[WIP] second stab at LDA/LSI
LDA still doesnt work quite as expected. Each word I've tried is
largely associated with a single topic. This may be because the data set
I use is too small, or because I used sentences as documents. It may be
hard to get the original document structure of wiki/gigaw because I seem
to have merged all documents into a single file, and then split that
into smaller file for concurrency issues.

This implementation also seems to do an unnecessary pass over the data